warehouse_env/pyproject.toml
[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "warehouse_env"
version = "0.1.0"
description = "A warehouse logistics environment for RL agents."
readme = "README.md"
authors = [{name = "XPOGBOY"}]
dependencies = [
    "openenv-core",
    "fastapi",
    "uvicorn",
    "pydantic",
    "python-multipart",
    "websockets",
    "openai"
]

[project.scripts]
server = "warehouse_env.server.app:main"

[project.optional-dependencies]
dev = ["pytest"]

[tool.setuptools.packages.find]
where = ["."]
include = ["warehouse_env*"]
